// src-ui/core/ipc/generated/chat/ChatArchive.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChatArchiveConversation } from "./ChatArchiveConversation";
import type { ChatArchiveMessage } from "./ChatArchiveMessage";

export type ChatArchive = { v: number, selfId: string, conversations: Array<ChatArchiveConversation>, messages: Array<ChatArchiveMessage>, };